+++ packages/schema/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @effect/schema
 
+## 0.57.2
+
+### Patch Changes
+
+- Updated dependencies [[`d5a1949`](https://github.com/Effect-TS/effect/commit/d5a19499aac7c1d147674a35ac69992177c7536c)]:
+  - effect@2.0.2
+
+## 0.57.1
+
+### Patch Changes
+
+- Updated dependencies [[`16bd87d`](https://github.com/Effect-TS/effect/commit/16bd87d32611b966dc42ea4fc979764f97a49071)]:
+  - effect@2.0.1
+
 ## 0.57.0
 
 ### Minor Changes